pyproject.toml
[project]
name = "proto-finder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "gitpython>=3.1.45",
    "mlxtend>=0.23.4",
    "pandas>=2.3.3",
    "tabulate>=0.9.0",
    "tdqm>=0.0.1",
    "tqdm>=4.67.1",
]
